examples/python/list_sql_tests/requirements.txt
--extra-index-url https://buf.build/gen/python 
getsynq-api-grpc-python
grpcio
grpcio-tools
requests
python-dotenv
